improved Pols

+++ tools/tbl/coxh3.jl
@@ -53,7 +53,7 @@ chevieset(:H3, :HeckeCharTable, function (param, sqrtparam)
                                         if IsInt(j[2])
                                             res = res * q ^ j[2]
                                         else
-                                            res = res * v ^ (2 * j[2])
+                                            res = res * v ^ Int(2 * j[2])
                                         end
                                         return res
                                     end, i)
